+++ src/bda/plone/checkout/interfaces.py
@@ -43,16 +43,25 @@ class IFieldsProvider(Interface):
         """
 
 
+class ICheckoutSettings(Interface):
+    """Settings for checkout.
+    """
+
+    def skip_payment(uid):
+        """Return flag whether to skip payment for data stored under uid.
+        """
+
+    def skip_payment_redirect_url(uid):
+        """URL to redirect if payment should be skipped for data stored under
+        uid.
+        """
+
+
 class ICheckoutAdapter(Interface):
     """Checkout persistence adapter.
     """
     vessel = Attribute(u"``zope.interface.mapping.IWriteMapping providing`` "
                        u"instance.")
-
-    skip_payment = Attribute(u"Flag whether to skip payment.")
-
-    skip_payment_redirect_url = Attribute(u"URL to redirect if payment should"
-                                          u" be skipped.")
 
     def save(providers, widget, data):
         """Save fields specific data.